[fix] #60 develop merge conflict fix
