+++ src/main/java/org/umc/travlocksserver/domain/vlock/exception/VlockCategoryException.java
@@ -1,6 +1,6 @@
 package org.umc.travlocksserver.domain.vlock.exception;
 
-import org.umc.travlocksserver.domain.vlock.constant.VlockCategoryErrorCode;
+import org.umc.travlocksserver.domain.vlock.code.VlockCategoryErrorCode;
 import org.umc.travlocksserver.global.exception.GeneralException;
 
 public class VlockCategoryException extends GeneralException {
