+++ src/main/java/org/umc/travlocksserver/domain/vlock/service/command/VlockCommandService.java
@@ -12,16 +12,16 @@ import org.umc.travlocksserver.domain.member.entity.Member;
 import org.umc.travlocksserver.domain.member.exception.MemberException;
 import org.umc.travlocksserver.domain.member.exception.code.MemberErrorCode;
 import org.umc.travlocksserver.domain.member.repository.MemberRepository;
-import org.umc.travlocksserver.domain.vlock.constant.VlockCategoryErrorCode;
-import org.umc.travlocksserver.domain.vlock.constant.VlockErrorCode;
-import org.umc.travlocksserver.domain.vlock.dto.vlock.VlockRequestDTO;
+import org.umc.travlocksserver.domain.vlock.code.VlockCategoryErrorCode;
+import org.umc.travlocksserver.domain.vlock.code.VlockErrorCode;
+import org.umc.travlocksserver.domain.vlock.dto.request.VlockRequestDTO;
+import org.umc.travlocksserver.domain.vlock.dto.response.VlockResponseDTO;
 import org.umc.travlocksserver.domain.vlock.entity.Vlock;
 import org.umc.travlocksserver.domain.vlock.entity.VlockCategory;
 import org.umc.travlocksserver.domain.vlock.exception.VlockCategoryException;
 import org.umc.travlocksserver.domain.vlock.exception.VlockException;
 import org.umc.travlocksserver.domain.vlock.repository.VlockCategoryRepository;
 import org.umc.travlocksserver.domain.vlock.repository.VlockRepository;
-import org.umc.travlocksserver.domain.vlock.service.VlockAsyncHandler;
 import org.umc.travlocksserver.domain.vlock.service.query.VlockCategoryQueryService;
 import org.umc.travlocksserver.infra.kakao.KakaoPlace;
 
@@ -32,13 +32,38 @@ import java.util.List;
 @Transactional
 public class VlockCommandService {
 
-    private final VlockRepository vlockRepository;
-    private final CityQueryService cityQueryService;
-    private final VlockCategoryQueryService vlockCategoryQueryService;
-    private final VlockCategoryRepository vlockCategoryRepository;
-    private final CityRepository cityRepository;
-    private final MemberRepository memberRepository;
-    private final VlockAsyncHandler vlockAsyncHandler;
+	private final MemberRepository memberRepository;
+	private final VlockCategoryRepository vlockCategoryRepository;
+	private final CityRepository cityRepository;
+	private final VlockRepository vlockRepository;
+	private final VlockCategoryQueryService vlockCategoryQueryService;
+	private final CityQueryService cityQueryService;
+
+	public VlockResponseDTO createVlock(Long memberId, VlockRequestDTO request) {
+		Member member = memberRepository.findById(memberId)
+			.orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
+
+		VlockCategory category = vlockCategoryRepository.findById(request.categoryId())
+			.orElseThrow(() -> new VlockException(VlockErrorCode.CATEGORY_NOT_FOUND));
+
+		City city = cityRepository.findWithRegionById(request.cityId())
+			.orElseThrow(() -> new CityException(CityErrorCode.CITY_NOT_FOUND));
+
+		Vlock newVlock = Vlock.create(
+			category,
+			city,
+			member,
+			request.name(),
+			request.latitude(),
+			request.longitude(),
+			request.address(),
+			request.memo()
+		);
+
+		Vlock savedVlock = vlockRepository.save(newVlock);
+
+		return VlockResponseDTO.from(savedVlock);
+	}
 
     // ⚪ 외부(카카오맵) API를 통해 블록을 삽입하는 메서드 (추천시 블록에 데이터가 너무 적을 경우 사용)
     public void upsertVlocksFromExternal(Long cityId, List<KakaoPlace> places) {
@@ -80,17 +105,4 @@ public class VlockCommandService {
                 .orElseGet(() -> vlockCategoryQueryService.getByName("기타")
                         .orElseThrow(() -> new VlockCategoryException(VlockCategoryErrorCode.DEFAULT_VLOCK_CATEGORY_NOT_FOUND)));
     };
-
-    public void createVlock(Long memberId, VlockRequestDTO request) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
-
-        VlockCategory category = vlockCategoryRepository.findById(request.categoryId())
-                .orElseThrow(() -> new VlockException(VlockErrorCode.CATEGORY_NOT_FOUND));
-
-        City city = cityRepository.findWithRegionById(request.cityId())
-                .orElseThrow(() -> new CityException(CityErrorCode.CITY_NOT_FOUND));
-
-        vlockAsyncHandler.saveVlockAsync(member, category, city, request);
-    }
 }
